// payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Page {
  id: number;
  title: string;
  slug?: string | null;
  hero?: {
    type?: ('videoBackground' | 'imageBackground' | 'none') | null;
    heading?: string | null;
    subheading?: string | null;
    videoUrl?: string | null;
    backgroundImage?: (number | null) | Media;
  };
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
// ...
export interface Header {
  id: number;
  logo?: (number | null) | Media;
  navItems?:
    | {
        link: HeaderNavItemLink;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeaderNavItemLink".
 */
export interface HeaderNavItemLink {
  label: string;
  type?: ('reference' | 'custom') | null;
  reference?: (number | null) | Page;
  url?: string | null;
  newTab?: boolean | null;
}